Refactor setup script

+++ setup.sh
@@ -1,40 +1,52 @@
 #!/bin/bash
 SETUP_DIR=$(dirname "$(readlink -f "$0")")
 LINK_ONLY=false
+INSTALL_PACKAGE=
 
+# Show usage
 _usage() {
   echo "$0 usage:" && grep " .)\\ #" $0 | sed -e 's/\(.\)) #/-\1/g'
   exit 0;
 }
 
+# Create symbolic link for config
+# $1: The source of symlink
+# $2: The destination of symlink
+# $3: When true, create parent directories to the link as needed. (default: true)
+# $4: Force removes the old destination, file or folder, and forces a new link (default: true)
 _link() {
+  local need_create_parent=${3:-true}
+  local is_force=${4:-true}
+
   if [[ -e "$2" || -L "$2" ]]; then
+  # Check destination is existed or is a symlink
     if test "$1" -ef "$(readlink -f "$2")"; then
+    # If symlink is setup same with source then finish
       echo "${2} linked"
       return
     else
-      rm -rf "$2"
-    fi
-  else
-    if [[ -d $1 && $HOME != "$(dirname "$2")" && ! -d "$(dirname "$2")" ]]; then
-      rm -rf "$(dirname "$2")"
+    # Delete destination if not a symlink or target of symlink is not same with source
+      [ "$is_force" = true ] && rm -rf "$2"
     fi
   fi
-  mkdir -p "$(dirname "$2")"
+  [ "$need_create_parent" = true ] && mkdir -p "$(dirname "$2")"
   ln -sv "$1" "$2"
 }
 
+# Create empty folder
+_create() {
+  mkdir -p "$*"
+}
+
 _install() {
-  if ! $LINK_ONLY; then
-    bash $SETUP_DIR/distros/install_linux_package.sh
-  fi
+  ! $LINK_ONLY && $SETUP_DIR/distros/install_linux_package.sh
 
   # Basic UI
   _link "$SETUP_DIR/X/resources" ~/.Xresources
   _link "$SETUP_DIR/X/font/files" ~/.local/share/fonts
   _link "$SETUP_DIR/X/font/config" ~/.config/fontconfig/conf.d
-  _link "$SETUP_DIR/X/theme/gtk" ~/.themes
-  _link "$SETUP_DIR/X/theme/icon-cursor" ~/.icons
+  _link "$SETUP_DIR/X/theme/gtk" ~/.local/share/themes
+  _link "$SETUP_DIR/X/theme/icon-cursor" ~/.local/share/icons
   _link "$SETUP_DIR/gtk/2/rc" ~/.gtkrc-2.0
   _link "$SETUP_DIR/gtk/3/settings.ini" ~/.config/gtk-3.0/settings.ini
   _link "$SETUP_DIR/qt/config" ~/.config/Trolltech.conf
@@ -47,13 +59,10 @@ _install() {
   _link "$SETUP_DIR/kitty" ~/.config/kitty
   _link "$SETUP_DIR/kitty/config" ~/.config/kitty/kitty.conf
   _link "$SETUP_DIR/fish" ~/.config/fish # Must setup fish shell before vim
-  if ! $LINK_ONLY; then
-    $SETUP_DIR/fish/setup.fish
-  fi
+  ! $LINK_ONLY && chsh -s $(which fish) && $SETUP_DIR/fish/setup.fish
   _link "$SETUP_DIR/vim" ~/.config/nvim
-  if ! $LINK_ONLY && command -v nvim > /dev/null 2>&1; then
-    nvim +PlugInstall +UpdateRemotePlugins +qa
-  fi
+  ! $LINK_ONLY && command -v nvim > /dev/null 2>&1 && nvim +PlugInstall +UpdateRemotePlugins +qa
+  ! $LINK_ONLY && bash $SETUP_DIR/bin/update.sh -n
 
   # Terminal application
   _link "$SETUP_DIR/tmux/config" ~/.tmux.conf
@@ -66,7 +75,7 @@ _install() {
 
   # Multimedia
   _link "$SETUP_DIR/mpd/config" ~/.config/mpd/mpd.conf
-  mkdir -p ~/.config/mpd/playlists
+  _create ~/.config/mpd/playlists
   _link "$SETUP_DIR/ncmpcpp" ~/.ncmpcpp
 
   # X11 miscelaneous
@@ -81,12 +90,10 @@ _install() {
   fi
 
   # GUI application
-  if ! $LINK_ONLY; then
-    $SETUP_DIR/firefox/setup.sh
-  fi
+  ! $LINK_ONLY && $SETUP_DIR/firefox/setup.sh
 }
 
-while getopts "hls" arg; do
+while getopts "hlsp" arg; do
   case $arg in
     s) # Update git submodule before setup
       git -C $SETUP_DIR submodule update --init --recursive --remote
